Added stepable
Necessary to teach an object how to behave line a pipeline of steps so any object that wants to have pipeable behavior can do so via the `#pipe` function.

Reek was updated to ignore the nested iterator and too many statement violations because -- when metaprogramming -- this is hard to avoid without making the code more complex by splitting things up further. At least all logic is self-contained as best as possible.